Limit organization name max length to 255 characters
Refactor OrgSelect tests [#186]
